IQL_KAS/requirements.txt
jax==0.4.30
jaxlib==0.4.30
flax==0.8.5
optax==0.2.4

numpy==1.26.4
scipy==1.13.1
pandas==2.3.3

gym==0.23.1
d4rl==1.1
mujoco-py==2.1.2.14

tqdm==4.67.3
tensorboardX==2.1

ml-collections==0.1.1
absl-py==2.3.1